+++ lib/src/generators/mocktail_fallback_values_generator.dart
@@ -2,7 +2,7 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:build/build.dart';
-import 'package:mockor/annotations.dart';
+import 'package:mockor/mockor.dart';
 import 'package:mockor/src/dartbuilders/mocktail_fallback_values_dart_builder.dart';
 import 'package:mockor/src/exceptions/mockor_exception.dart';
 import 'package:mockor/src/generators/common/resolve_asset_uri.dart';
@@ -115,10 +115,10 @@ extension on List<ResolvedType> {
 
   void _makeUnique(
       ResolvedType resolvedType, Map<String, ResolvedType> uniqueEntries) {
-    if (uniqueEntries.containsKey(resolvedType.displayNameUnique)) {
+    if (uniqueEntries.containsKey(resolvedType.nameUnique)) {
       _makeUnique(_appendDollar(resolvedType), uniqueEntries);
     } else {
-      uniqueEntries[resolvedType.displayNameUnique] = resolvedType;
+      uniqueEntries[resolvedType.nameUnique] = resolvedType;
     }
   }