+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1
+
+- Improve package description
+- Move builder.dart to src folder.
+- Fix compatibility with analyzer 1.7.0
+- Change annotations package to mockor.
+
 ## 1.0.0
 
 - Initial release